[TEMPORARY] Re-point A/B testing preference for test
Instead of writing to the visualeditor-enable preference (which is
user-facing), write to the visualeditor-testabpreference one (which
is made up and won't impact anything), so that we can test that the
A/B bucketing code works as expected.

This patch will be reverted once bucketing is confirmed to be working.

Change-Id: Ia738146c5bf68795bf812d64847ba74498888cd5
(cherry-picked from 1a42383cdd6255264100c9670e044d8aeeca27f6)